introduce initialise.py
